changed type to double

+++ Module1.TypesAndClasses/Module1.TypesAndClasses/Interfaces/IShape.cs
@@ -2,8 +2,8 @@
 {
     public interface IShape
     {
-        long Square();
+        double Square();
 
-        int Perimeter();
+        double Perimeter();
     }
 }
